requirements.txt
haystack-ai>=2.0.0
trafilatura>=1.6.0
lxml>=5.1.0
python-dotenv>=1.0.0
openai>=1.0.0
unstructured>=0.13.0
nltk>=3.8.1
sentence-transformers>=2.5.1
